chore: move camera activity from utils to ui
