+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <uses-feature android:name="android.hardware.camera" />
+
     <uses-permission android:name="android.permission.CAMERA" />
     <uses-permission
         android:name="android.permission.WRITE_EXTERNAL_STORAGE"
@@ -20,8 +21,8 @@
         tools:targetApi="31">
 
         <activity
-            android:name="com.nudriin.fits.utils.CameraActivity"
-            android:exported="false"/>
+            android:name="com.nudriin.fits.ui.camera.CameraActivity"
+            android:exported="false" />
 
         <activity
             android:name=".MainActivity"
